+++ src/js/core/height-viewport.js
@@ -1,4 +1,4 @@
-import {css, endsWith, height, /* isIE, */ isNumeric, isString, offset, query, toFloat} from 'uikit-util';
+import {css, endsWith, height, isIE, isNumeric, isString, offset, query, toFloat} from 'uikit-util';
 
 export default {
 
@@ -76,9 +76,11 @@ export default {
             }
 
             // IE 11 fix (min-height on a flex container won't apply to its flex items)
-            let height;
-            if (/* isIE && */(height = Math.round(toFloat(css(this.$el, 'minHeight')))) >= offsetHeight(this.$el)) {
-                css(this.$el, 'height', height);
+            if (isIE) {
+                const height = Math.round(toFloat(css(this.$el, 'minHeight')));
+                if (height >= offsetHeight(this.$el)) {
+                    css(this.$el, 'height', height);
+                }
             }
 
         },